Text only over background option

+++ Include/GeneralGoodImages.h
@@ -19,12 +19,12 @@
 	#include <3ds/svc.h>
 #endif
 
-void drawTexturePartScaleRotate(CrossTexture* texture, int x, int y, float tex_x, float tex_y, float tex_w, float tex_h, float x_scale, float y_scale, float rad);
-void drawTexturePartScaleTint(CrossTexture* passedTexture, int destX, int destY, int texX, int texY, int texW, int texH, float texXScale, float texYScale, unsigned char r, unsigned char g, unsigned b);
-void drawTexturePartScale(CrossTexture* passedTexture, int destX, int destY, int texX, int texY, int texW, int texH, float texXScale, float texYScale);
-void drawTextureScaleSize(CrossTexture* passedTexture, int destX, int destY, float texXScale, float texYScale);
-void drawTextureScaleTint(CrossTexture* passedTexture, int destX, int destY, float texXScale, float texYScale, unsigned char r, unsigned char g, unsigned char b);
-void drawTextureScale(CrossTexture* passedTexture, int destX, int destY, float texXScale, float texYScale);
+void drawTexturePartScaleRotate(CrossTexture* texture, int x, int y, int tex_x, int tex_y, int tex_w, int tex_h, double x_scale, double y_scale, double rad);
+void drawTexturePartScaleTint(CrossTexture* passedTexture, int destX, int destY, int texX, int texY, int texW, int texH, double texXScale, double texYScale, unsigned char r, unsigned char g, unsigned b);
+void drawTexturePartScale(CrossTexture* passedTexture, int destX, int destY, int texX, int texY, int texW, int texH, double texXScale, double texYScale);
+void drawTextureScaleSize(CrossTexture* passedTexture, int destX, int destY, double texXScale, double texYScale);
+void drawTextureScaleTint(CrossTexture* passedTexture, int destX, int destY, double texXScale, double texYScale, unsigned char r, unsigned char g, unsigned char b);
+void drawTextureScale(CrossTexture* passedTexture, int destX, int destY, double texXScale, double texYScale);
 void drawTexture(CrossTexture* passedTexture, int _destX, int _destY);
 void freeTexture(CrossTexture* passedTexture);
 int getTextureHeight(CrossTexture* passedTexture);
@@ -34,6 +34,6 @@ CrossTexture* loadPNG(char* path);
 CrossTexture* loadPNGBuffer(void* _passedBuffer, int _passedBufferSize);
 CrossTexture* loadJPGBuffer(void* _passedBuffer, int _passedBufferSize);
 void drawTextureAlpha(CrossTexture* passedTexture, int _destX, int _destY, unsigned char alpha);
-void drawTextureScaleAlpha(CrossTexture* passedTexture, int destX, int destY, float texXScale, float texYScale, unsigned char alpha);
+void drawTextureScaleAlpha(CrossTexture* passedTexture, int destX, int destY, double texXScale, double texYScale, unsigned char alpha);
 
 #endif /* GENERALGOODGRAPHICS_H */
